+++ pyoffers/api.py
@@ -87,10 +87,13 @@ class HasOffersAPI:
         return self._session
 
     @retry
-    def _call(self, target, method, single_result=True, raw=False, files=None, **kwargs):
+    def _call(self, target, method, target_class=None, single_result=True, raw=False, files=None, **kwargs):
         """
         Low-level call to HasOffers API.
+        :param target_class: type of resulting object/objects.
         """
+        if target_class is None:
+            target_class = target
         params = prepare_query_params(
             NetworkToken=self.network_token,
             NetworkId=self.network_id,
@@ -108,7 +111,7 @@ class HasOffersAPI:
         self.logger.debug('Response [%s]: %s', response.status_code, response.text)
         response.raise_for_status()
         data = response.json(object_pairs_hook=OrderedDict)
-        return self.handle_response(data, target=target, single_result=single_result, raw=raw)
+        return self.handle_response(data, target=target_class, single_result=single_result, raw=raw)
 
     def handle_response(self, content, target=None, single_result=True, raw=False):
         """